mk-sd-card.sh: Use process local mount points to allow parallel execution
The usage of a global mount directory prevents the script from being
run in parallel for multiple cards.  Changing the mount point to a
directory in /tmp and including the PID of the shell solves this
problem.

Tests suggest that mkfs is locked against itself even on different
devices, so a full parallelism is not achieved.

Signed-off-by: Detlev Zundel <detlev.zundel@ebv.com>