mk-sd-card.sh, environment.txt: Add bitstream for SoCrates v2 and use it
The pinout for the FPGA Leds has changed on SoCrates v2 hardware.  As
we cannot determine from software what version we are running on, we
have to leave a minimal adjustment up to the user.  Two bitstreams for
the traditional SysID / PIO demo are provided for both board
versions.  Adjusting the environment variable "fpga_image" allows to
switch between them.

Signed-off-by: Detlev Zundel <detlev.zundel@ebv.com>